Feat: Add Admin notice page and apis

+++ src/app/api/editor/deletePost/route.ts
@@ -18,13 +18,22 @@ export async function DELETE(request: Request) {
   })
 
   if (post && user && (post.user_id === user.user_id || user.grade_id === 1)) {
-    const delete_post = await prisma.post.delete({
-      where: {
-        post_id,
-      },
-    })
+    let result: any
 
-    if (!delete_post) return new Response(null, { status: 500 })
+    if (post.board_id === 13) {
+      result = await prisma.post.update({
+        where: { post_id },
+        data: { user_id: '(알 수 없음)' },
+      })
+    } else {
+      result = await prisma.post.delete({
+        where: {
+          post_id,
+        },
+      })
+    }
+
+    if (!result) return new Response(null, { status: 500 })
     else return new Response(null, { status: 200 })
   } else return new Response(null, { status: 401 })
 }
